Keep memory_profiler in single file, drop support of python <3.4
